Extract unit rules into parser
