+++ src/value/mod.rs
@@ -10,11 +10,12 @@ use functions::get_builtin_function;
 use nom::multispace;
 use num_rational::Rational;
 use num_traits::{One, Signed, Zero};
-use parser::call_args;
+use parser::formalargs::call_args;
+use parser::unit::unit;
 use parser::util::{is_name_char, name, opt_spacelike, spacelike2};
 use std::fmt;
 use std::str::{FromStr, from_utf8};
-use unit::{Unit, unit};
+use unit::Unit;
 use variablescope::Scope;
 
 /// A sass value.
